refactor(card): relocate Cards to card folder
