+++ apps/ui/src/components/cards/ApplicationCard.tsx
